moving testsuite console log to helpers.js
